+++ Tests/RulesTests+Syntax.swift
@@ -621,12 +621,24 @@ class SyntaxTests: RulesTests {
     }
 
     func testEnumNamespacesImportClass() {
-        let input = "import class MyUIKit.AutoHeightTableView"
+        let input = """
+        import class MyUIKit.AutoHeightTableView
+
+        enum Foo {
+            static var bar: String
+        }
+        """
         testFormatting(for: input, rule: FormatRules.enumNamespaces)
     }
 
     func testEnumNamespacesImportStruct() {
-        let input = "import struct Core.CurrencyFormatter"
+        let input = """
+        import struct Core.CurrencyFormatter
+
+        enum Foo {
+            static var bar: String
+        }
+        """
         testFormatting(for: input, rule: FormatRules.enumNamespaces)
     }
 
@@ -926,6 +938,56 @@ class SyntaxTests: RulesTests {
         """
         let options = FormatOptions(enumNamespaces: .structsOnly)
         testFormatting(for: input, rule: FormatRules.enumNamespaces, options: options)
+    }
+
+    func testEnumNamespacesAfterImport() {
+        // https://github.com/nicklockwood/SwiftFormat/issues/1569
+        let input = """
+        import Foundation
+
+        final class MyViewModel2 {
+            static let = "A"
+        }
+        """
+
+        let output = """
+        import Foundation
+
+        enum MyViewModel2 {
+            static let = "A"
+        }
+        """
+
+        testFormatting(for: input, output, rule: FormatRules.enumNamespaces)
+    }
+
+    func testEnumNamespacesAfterImport2() {
+        // https://github.com/nicklockwood/SwiftFormat/issues/1569
+        let input = """
+        final class MyViewModel {
+            static let = "A"
+        }
+
+        import Foundation
+
+        final class MyViewModel2 {
+            static let = "A"
+        }
+        """
+
+        let output = """
+        enum MyViewModel {
+            static let = "A"
+        }
+
+        import Foundation
+
+        enum MyViewModel2 {
+            static let = "A"
+        }
+        """
+
+        testFormatting(for: input, output, rule: FormatRules.enumNamespaces)
     }
 
     // MARK: - numberFormatting
@@ -3676,5 +3738,139 @@ class SyntaxTests: RulesTests {
 
         let options = FormatOptions(swiftVersion: "5.9")
         testFormatting(for: input, rule: FormatRules.conditionalAssignment, options: options)
+    }
+
+    func testDoesntConvertConditionalCastInSwift5_9() {
+        // The following code doesn't compile in Swift 5.9 due to this issue:
+        // https://github.com/apple/swift/issues/68764
+        //
+        //  let result = if condition {
+        //    foo as? String
+        //  } else {
+        //    "bar"
+        //  }
+        //
+        let input = """
+        let result1: String?
+        if condition {
+            result1 = foo as? String
+        } else {
+            result1 = "bar"
+        }
+
+        let result2: String?
+        switch condition {
+        case true:
+            result2 = foo as? String
+        case false:
+            result2 = "bar"
+        }
+        """
+
+        let options = FormatOptions(swiftVersion: "5.9")
+        testFormatting(for: input, rule: FormatRules.conditionalAssignment, options: options)
+    }
+
+    func testAllowsAsWithinInnerScope() {
+        let input = """
+        let result: String?
+        switch condition {
+        case true:
+            result = method(string: foo as? String)
+        case false:
+            result = "bar"
+        }
+        """
+
+        let output = """
+        let result: String? = switch condition {
+        case true:
+            method(string: foo as? String)
+        case false:
+            "bar"
+        }
+        """
+
+        let options = FormatOptions(swiftVersion: "5.9")
+        testFormatting(for: input, output, rule: FormatRules.conditionalAssignment, options: options)
+    }
+
+    // TODO: update branches parser to handle this case properly
+    func testIgnoreSwitchWithConditionalCompilation() {
+        let input = """
+        func foo() -> String? {
+            let result: String?
+            switch condition {
+            #if os(macOS)
+            case .foo:
+                result = method(string: foo as? String)
+            #endif
+            case .bar:
+                return nil
+            }
+            return result
+        }
+        """
+
+        let options = FormatOptions(ifdefIndent: .noIndent, swiftVersion: "5.9")
+        testFormatting(for: input, rule: FormatRules.conditionalAssignment, options: options)
+    }
+
+    // TODO: update branches parser to handle this scenario properly
+    func testIgnoreSwitchWithConditionalCompilation2() {
+        let input = """
+        func foo() -> String? {
+            let result: String?
+            switch condition {
+            case .foo:
+                result = method(string: foo as? String)
+            #if os(macOS)
+            case .bar:
+                return nil
+            #endif
+            }
+            return result
+        }
+        """
+
+        let options = FormatOptions(ifdefIndent: .noIndent, swiftVersion: "5.9")
+        testFormatting(for: input, rule: FormatRules.conditionalAssignment, options: options)
+    }
+
+    func testConvertsConditionalCastInSwift5_10() {
+        let input = """
+        let result1: String?
+        if condition {
+            result1 = foo as? String
+        } else {
+            result1 = "bar"
+        }
+
+        let result2: String?
+        switch condition {
+        case true:
+            result2 = foo as? String
+        case false:
+            result2 = "bar"
+        }
+        """
+
+        let output = """
+        let result1: String? = if condition {
+            foo as? String
+        } else {
+            "bar"
+        }
+
+        let result2: String? = switch condition {
+        case true:
+            foo as? String
+        case false:
+            "bar"
+        }
+        """
+
+        let options = FormatOptions(swiftVersion: "5.10")
+        testFormatting(for: input, output, rule: FormatRules.conditionalAssignment, options: options)
     }
 }